Prd import base

See merge request apps/console!69

+++ src/management/organization.py
@@ -159,11 +159,8 @@ def register_noti_key(platform):
       return render_template("register_android.html", referrer=referrer,
                              product_list=product_list)
   else:
-    # TODO: : Comment out about Not supported feature from mib cloud
-    #allow_dict = json.loads(request.form['allowDict'])
-    #allow_model_id_list = json.loads(request.form['allowModelIdList'])
-    allow_dict = {}
-    allow_model_id_list = []
+    allow_dict = json.loads(request.form['allowDict'])
+    allow_model_id_list = json.loads(request.form['allowModelIdList'])
     if platform == "ios":
       bundle_id = request.form['bundleId']
       password = request.form['password']
@@ -252,31 +249,32 @@ def update_noti_key(noti_key_id):
                             allow_model_id_list=allow_model_id_list,
                             allow_dict=allow_dict)
   else:
-    # TODO: : Comment out about Not supported feature from mib cloud
-    #allow_dict = json.loads(request.form['allowDict'])
-    #allow_model_id_list = json.loads(request.form['allowModelIdList'])
-    allow_dict = {}
-    allow_model_id_list = []
-    noti_key = in_apis.get_noti_key(noti_key_id)
-    ret = apis.update_allow_noti_key(current_user.organization_id, noti_key.name, allow_dict)
-    if ret:
-      permit_list = in_apis.get_noti_model_permission_list_by_noti_id(noti_key.id)
-      if permit_list:
-        for permit in permit_list:
-          if permit.model_id in allow_model_id_list:
-            in_apis.update_noti_model_permission(permit.id, current_user.email,
-                                                noti_key.id, permit.model_id)
-            allow_model_id_list.remove(permit.model_id)
-          else:
-            in_apis.delete_noti_model_permission(permit.id)
-      for model_id in allow_model_id_list:
-        in_apis.create_noti_model_permission(current_user.email, noti_key.id,
+    allow_dict = json.loads(request.form['allowDict'])
+    allow_model_id_list = json.loads(request.form['allowModelIdList'])
+    if allow_dict and allow_model_id_list:
+      noti_key = in_apis.get_noti_key(noti_key_id)
+      ret = apis.update_allow_noti_key(current_user.organization_id, noti_key.name,
+                                       allow_dict)
+      if ret:
+        permit_list = in_apis.get_noti_model_permission_list_by_noti_id(noti_key.id)
+        if permit_list:
+          for permit in permit_list:
+            if permit.model_id in allow_model_id_list:
+              in_apis.update_noti_model_permission(permit.id, current_user.email,
+                                                  noti_key.id, permit.model_id)
+              allow_model_id_list.remove(permit.model_id)
+            else:
+              in_apis.delete_noti_model_permission(permit.id)
+        for model_id in allow_model_id_list:
+          in_apis.create_noti_model_permission(current_user.email, noti_key.id,
                                               model_id)
-      return redirect('/management/organization/notification')
+        return redirect('/management/organization/notification')
+      else:
+        logging.warning("Failed to update noti key. id : %s, user : %s, message : %s",
+                        noti_key_id, current_user.email, ret)
+        abort(500)
     else:
-      logging.warning("Failed to update noti key. id : %s, user : %s, message : %s",
-                       noti_key_id, current_user.email, ret)
-      abort(500)
+      return redirect('/management/organization/notification')
 
 
 def delete_noti_key(noti_key_id):
